fix: moved the logic to the scaler class, the changes required to have these checks in Fittable/Invertible were too heavy considering the diversity in data-transformer (static_covariate_transformer being the one challenging the logic)
